requirements.txt
torch~=2.0.1
numpy~=1.24.3
opencv-python~=4.8.0.74
keras~=2.13.1
tqdm~=4.65.0
Pillow~=10.0.0
PyYAML~=6.0.1
huggingface_hub
tensorflow
realesrgan
realcugan-ncnn-py